refactor: rename models/snek -> models/naff
