+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,5 +1,5 @@
 blank_issues_enabled: false
 contact_links:
-  - name: 💬 dis-snek discord
+  - name: 💬 dis-naff discord
     url: https://discord.gg/NTSHu97tHg
-    about: Join the dis-snek discord server!
+    about: Join the dis-naff discord server!
