+++ mkdocs.yml
@@ -39,7 +39,7 @@ theme:
 extra:
   social:
     - icon: fontawesome/brands/discord
-      link: https://discord.gg/dis-snek
+      link: https://discord.gg/dis-naff
       name: Discord Server
     - icon: fontawesome/brands/github
       link: https://github.com/Discord-Snake-Pit/Dis-Snek
