+++ naff/models/naff/protocols.py
@@ -4,7 +4,7 @@ from typing import Protocol, Any, TYPE_CHECKING
 from naff.client.const import T_co
 
 if TYPE_CHECKING:
-    from naff.models.snek.context import Context
+    from naff.models.naff.context import Context
 
 __all__ = ("Converter",)
